==> Week_4/1.2.EnergyClasses_Consolidate.ipynb <==
## Energy Label Classification

### Consolidating the collected information - Part 1

import utils
from datetime import datetime
import time

import os
import ast
import json
import requests
import multiprocessing as mp

import pandas as pd

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
pd.set_option('display.max_colwidth', 500)

Define: 
- the file paths for the files containing the unprocessed HTTP responses
- the output file where we will be saving the cleaned data

mydir = os.path.join(os.getcwd(), 'data', 'new', 'data')
files = [os.path.join(mydir, f) for f in os.listdir(mydir) if os.path.isfile(os.path.join(mydir, f))]
features_f = os.path.join(os.getcwd(), 'data', 'energy_classes.csv')

#### Parse the JSON files and store the results in a CSV file fo further processing

start = datetime.now()
print(start)
for j in range(len(files)):    
    print('[{}]Parsing {}'.format(datetime.now(), files[j]))
    contents = open(files[j], "r").read() 
    rows = ['{'+item+'}' for item in contents.split('}{')]
    rows[0] = rows[0][1:]
    rows[-1] = rows[-1][:-1]

    df = pd.DataFrame()

    for i in range(len(rows)):
        rows[i] = rows[i].replace('"', '$')
        rows[i] = rows[i].replace("'", '~')
        rows[i] = rows[i].replace('$', "'")
        rows[i] = rows[i].replace('~', '"')
        rows[i] = rows[i].replace('datetime.datetime(', '"datetime.datetime(')
        rows[i] = rows[i].replace('),', ')",')

        try:
            d = ast.literal_eval(rows[i])
            newdf = pd.io.json.json_normalize(d)
            df = df.append(newdf, ignore_index=True)
        except Exception as e:
            print(i)
            print(rows[i])
            print(str(e))
    try:
        df['data.features'].to_csv(features_f, mode='a', header=False)
    except Exception as e:
        print(j)
        print(str(e))
        
print(datetime.now()-start)

df.columns

#### How does the data that we are interested in actually look like?

feat = df['data.features']

feat
