==> Week_4/1.3.EDA_energy_classes.ipynb <==
## Energy Label Classification

### Consolidating the collected information - Part 2

We saw previously that we are mainly interested in the __features__ column from the CSV we created out of the HTT requests. 

Look into the respective column more closely and make sure we're discarding as much junk as possible.

from datetime import datetime
import time

import multiprocessing as mp
import os
import ast
import json

import utils

import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
pd.set_option('display.max_colwidth', 500)
import pandas_explode 
pandas_explode.patch() # adds a `df.explode` method to all DataFrames 

energy_cls_file = os.path.join(os.getcwd(), 'data', 'energy_classes.csv')
energy_cls = pd.read_csv(energy_cls_file)

energy_cls.columns = ['initial_index', 'features']
energy_cls.tail(1)

energy_cls.dtypes

energy_cls = energy_cls[energy_cls['features']!='[]']
energy_cls.shape

energy_cls['features_len'] = energy_cls['features'].apply(lambda x: len(str(x)))
energy_cls = energy_cls[energy_cls['features_len']>4]
energy_cls.shape

energy_cls['features'] = energy_cls['features'].apply(lambda x: ast.literal_eval(x))

start = datetime.now()
energy_cls.drop(columns=['features_len'], inplace=True)
energy_cls_new = energy_cls.explode('features')
print(datetime.now()-start)

energy_cls_new.head()

Extract the __geometry__ piece of information from the __features__ column.

energy_cls_new['geometry'] = energy_cls_new['features'].apply(lambda x: x['geometry'])
energy_cls_new.head(1)

#### Save the results in a file for further processing

energy_cls_file_clean = os.path.join(os.getcwd(), 'data', 'energy_classes_clean.csv')
energy_cls_new.to_csv(energy_cls_file_clean, sep=';', index=False)

Double check the results.

os.path.join(os.getcwd(), 'data', 'energy_classes_clean.csv')

!head -n 2 '/home/osboxes/courses/ibm/data/energy_classes_clean.csv'
