==> Week_4/2.2.EDA_solar.ipynb <==
## Solar Potential
### Consolidating the collected information - Part 2

The original data source for Solar Potential had the coordinates in Easting, Northing coordinates so we need to transform them to  Latitude and Longitude such that we can link them to the Energy Label Classification information (also such that we can better relate to them in general).

from datetime import datetime
import time

import multiprocessing as mp
import os
import ast
import json

import utm
import utils

import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)
pd.set_option('display.max_colwidth', 500)

Define: 
- the file paths for the files containing the solar potential information
- the output file where we will be saving the information after we have transformed the coordinates from Eastin, Northing to Latitude and Longitude

mydir = os.path.join(os.getcwd(), 'data', 'solar_clean')
files = [os.path.join(mydir, f) for f in os.listdir(mydir) if os.path.isfile(os.path.join(mydir, f))]
files = sorted(files)
print('We have {} files'.format(len(files)))

new_dir = os.path.join(os.getcwd(), 'data', 'solar_ll')
if not os.path.exists(new_dir):
    os.makedirs(new_dir)
    print('Created {}'.format(new_dir))
else:
    print('Dir {} already exists'.format(new_dir))

def to_latlong(east_north):
    east = east_north[0]
    north = east_north[1]
    try:
        lat_long = list(utm.to_latlon(east,north,32,'N'))
        return lat_long
    except Exception as e:
        print('Error converting {} -- {}'.format(east_north, str(e)))
        return None

def transform_geom(row):
    geom = row['geometry']
    geom['coordinates'] = [to_latlong(point[0:2]) for point in geom['coordinates'][0]]
    return geom

files = sorted(files)
test_index = [0]

for i in test_index:
    start = datetime.now()
    print(start)
    with open(files[i]) as f:
        df = pd.read_csv(f, sep=';')
        df['geometry_ll'] = df.apply(lambda x: transform_geom(x), axis=1)
        print(display(df.head(1)))
        print(datetime.now()-start)
        print(datetime.now())        

pool = mp.Pool(mp.cpu_count())
results = pool.map(utils.add_latlong, [(files[i], new_dir) for i in range(len(files))])

#Test one file
df = pd.read_csv(results[0])
df.head(1)
